requirements.txt
colorama>=0.4.6
Pillow>=9.4.0